CLDR-14615 Modernize by fixing UI glitches: Prev/Next buttons, etc.
-Fix progress bar (Votes/Abstain)

-Fix highlighting of Prev/Next/Toggle buttons

-Fix Priority Items Summary missing from gear menu (perm.userCanUseVettingSummary)

-Improve various alignments

-Use webpack to bundle the forum css

-Update pretty.sh for new esm, css locations

-Move setToptitleVisibility to cldrGui
