+++ tools/cldr-apps/js-unittest/pretty.sh
@@ -1,5 +1,6 @@
-npx prettier --write tools/cldr-apps/src/main/webapp/js/esm/*.js
+npx prettier --write tools/cldr-apps/js/src/*.js
+npx prettier --write tools/cldr-apps/js/src/css/*.css
+npx prettier --write tools/cldr-apps/js/src/esm/*.js
+npx prettier --write tools/cldr-apps/js/src/views/*.vue
 npx prettier --write tools/cldr-apps/js-unittest/*.js
 npx prettier --write tools/cldr-apps/js-unittest/*.html
-npx prettier --write tools/cldr-apps/js/src/*.js
-npx prettier --write tools/cldr-apps/js/src/views/*.vue
