+++ tools/cldr-apps/src/main/webapp/css/redesign.css
@@ -510,13 +510,6 @@ h3.collapse-review > span:first-child {position: relative; top: -220px;display: 
     margin-right:  3px;
 }
 
-
-div#additional-top {
-	position: fixed;
-	left: calc(50% - 250px);
-	top:140px;
-}
-
 div#additional-top div.topnotices {
     position:  fixed;
     left:  0;
@@ -1035,7 +1028,7 @@ html {
 
 .sidebyside-column-top {
 	flex: none;
-	margin: 4px;
+	margin: 0;
 }
 
 .sidebyside-scrollable {
@@ -1049,15 +1042,12 @@ html {
 
 #toptitle {
 	font-size: medium;
-}
-
-#toparea #toptitle {
+	display: none;
 	padding-top: 0;
-	margin-top: 0;
-	margin-bottom: 0;
-	line-height: 2.5em;
-	font-weight: normal;
 	padding-bottom: 0;
+	margin-top: 0.25em;
+	margin-bottom: 0.25em;
+	margin-left: 1ex;
 }
 
 #toptitle span.title-cldr {
@@ -1067,6 +1057,7 @@ html {
 #toptitle h1 {
 	display: inline-block;
 	font-size: 28px;
+	margin: 0;
 }
 
 #toptitle h1::before {
@@ -1079,8 +1070,8 @@ html {
 
 #itemInfo div.xpath {
 	position: fixed;
-	/* TODO: get the right width of the "slider" for left sidebar; cf. beware-left-sidebar */
-	left: 34px;
+	/* The left sidebar "slider" width is 2% ( = 20% x 10%); cf. beware-left-sidebar */
+	left: 2.1%;
 	bottom: 0.1em;
 	color: #a19a9a;
 	padding:  2px;
@@ -1097,24 +1088,40 @@ html {
 }
 
 .beware-left-sidebar {
-	/* TODO: get the right width of the "slider" for left sidebar; cf. #itemInfo div.xpath */
-	padding-left: 34px;
+	/* The left sidebar "slider" width is 2% ( = 20% x 10%); cf. #itemInfo div.xpath */
+	padding-left: 2.1%;
+}
+
+#nav-page {
+	width: 100%;
+	display: flex;
+	flex-direction: row;
+	flex-wrap: wrap;
+	justify-content: space-between;
+	align-items: center;
+	margin: 0.25em 0.25em 0 0;
+	padding: 0;
+	box-sizing: border-box;
+}
+
+.cldr-nav-btn {
+	padding: 1px 5px;
+	margin-left: 0;
+	margin-right: 1ex;
+	font-size: 12px;
+	line-height: 1.5;
+	border-radius: 3px;
+}
+
+.toggle-right {
+	margin-left: 3em;
+	margin-right: 2ex;
 }
 
 .nav-progress {
 	width: 150px;
-	margin-left: 15px;
-}
-
-#nav-page {
-	align-items: center;
-	width: 100%;
-	display: flex;
-	flex-wrap: wrap;
-	justify-content: space-between;
-	text-align: center;
-	align-items: center;
-	margin: 0;
-	padding: 0em 1em 0em 1em;
-	box-sizing: border-box;
+	height: 20px;
+	margin: 0 8px 4px 0;
+	display: inline-block;
+	vertical-align: middle;
 }
